chore(claim): exctract gas claim into shared function
