+++ src/browser/actions/makeClaimActions.js
@@ -1,25 +1,10 @@
 import { createActions } from 'spunky';
-import { api } from '@cityofzion/neon-js';
+
+import claimGas from 'shared/util/claimGas';
 
 import generateDAppActionId from './generateDAppActionId';
 
 export const ID = 'claim';
-
-const claimGas = async ({ net, address, wif }) => {
-  const config = {
-    net,
-    address,
-    privateKey: wif
-  };
-
-  const { response: { result, txid } } = await api.claimGas(config, api.neoscan);
-
-  if (!result) {
-    throw new Error('Claim failed.');
-  }
-
-  return txid;
-};
 
 export default function makeClaimActions(sessionId, requestId) {
   const id = generateDAppActionId(sessionId, `${ID}-${requestId}`);
