+++ src/shared/actions/claimActions.js
@@ -1,23 +1,8 @@
 import { createActions } from 'spunky';
-import { api } from '@cityofzion/neon-js';
+
+import claimGas from '../util/claimGas';
 
 export const ID = 'claim';
-
-const claimGas = async ({ net, address, wif }) => {
-  const config = {
-    net,
-    address,
-    privateKey: wif
-  };
-
-  const { response: { result, txid } } = await api.claimGas(config, api.neoscan);
-
-  if (!result) {
-    throw new Error('Claim failed.');
-  }
-
-  return txid;
-};
 
 export default createActions(ID, ({ net, address, wif }) => () => {
   return claimGas({ net, address, wif });
